refactor: explore new animations
