+++ apps/web/modules/onboarding/components/OnboardingLayout.tsx
@@ -1,9 +1,7 @@
 "use client";
 
 import classNames from "classnames";
-import { AnimatePresence, motion } from "framer-motion";
 import { signOut } from "next-auth/react";
-import { usePathname } from "next/navigation";
 import { Children, type ReactNode } from "react";
 
 import { useLocale } from "@calcom/lib/hooks/useLocale";
@@ -19,7 +17,6 @@ type OnboardingLayoutProps = {
 
 export const OnboardingLayout = ({ userEmail, currentStep, totalSteps, children }: OnboardingLayoutProps) => {
   const { t } = useLocale();
-  const pathname = usePathname();
 
   const handleSignOut = () => {
     signOut({ callbackUrl: "/auth/logout" });
@@ -30,49 +27,21 @@ export const OnboardingLayout = ({ userEmail, currentStep, totalSteps, children 
   const column1 = childrenArray[0];
   const column2 = childrenArray[1];
 
-  // Animation variants for entry and exit
-  const containerVariants = {
-    initial: {
-      opacity: 0,
-      y: 20,
-    },
-    animate: {
-      opacity: 1,
-      y: 0,
-    },
-    exit: {
-      opacity: 0,
-      y: -20,
-    },
-  };
-
   return (
     <div className="bg-muted flex min-h-screen w-full flex-col items-center justify-between overflow-clip rounded-[12px] px-6 py-10">
       {/* Logo and container - centered */}
       <div className="flex w-full flex-1 flex-col items-center justify-center gap-8">
         <Logo className="h-5 w-auto shrink-0" />
-        <AnimatePresence mode="wait">
-          <motion.div
-            key={pathname}
-            className="border-subtle bg-default grid w-full max-w-[532px] grid-cols-1 gap-6 overflow-hidden rounded-2xl border px-12 py-10 xl:h-[690px] xl:max-w-[1130px] xl:grid-cols-[40%_1fr] xl:pl-10 xl:pr-0"
-            variants={containerVariants}
-            initial="initial"
-            animate="animate"
-            exit="exit"
-            transition={{
-              duration: 0.5,
-              ease: "easeInOut",
-            }}>
-            {/* Column 1 - Always visible, 40% on xl+ */}
-            <div className="flex min-h-0 flex-col">{column1}</div>
-            {/* Column 2 - Hidden on mobile, visible on xl+, 60% on xl+ */}
-            {column2 && (
-              <div className="hidden h-full max-h-full min-h-0 flex-col overflow-hidden xl:flex">
-                {column2}
-              </div>
-            )}
-          </motion.div>
-        </AnimatePresence>
+        <div className="border-subtle bg-default grid w-full max-w-[532px] grid-cols-1 gap-6 overflow-hidden rounded-2xl border px-12 py-10 xl:h-[690px] xl:max-w-[1130px] xl:grid-cols-[40%_1fr] xl:pl-10 xl:pr-0">
+          {/* Column 1 - Always visible, 40% on xl+ */}
+          <div className="flex min-h-0 flex-col">{column1}</div>
+          {/* Column 2 - Hidden on mobile, visible on xl+, 60% on xl+ */}
+          {column2 && (
+            <div className="hidden h-full max-h-full min-h-0 flex-col overflow-hidden xl:flex">
+              {column2}
+            </div>
+          )}
+        </div>
       </div>
 
       {/* Footer with progress dots and sign out */}
