+++ apps/web/modules/onboarding/components/onboarding-organization-browser-view.tsx
@@ -1,6 +1,8 @@
 "use client";
 
 import classNames from "classnames";
+import { AnimatePresence, motion } from "framer-motion";
+import { usePathname } from "next/navigation";
 
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
@@ -24,8 +26,25 @@ export const OnboardingOrganizationBrowserView = ({
   bannerUrl,
 }: OnboardingOrganizationBrowserViewProps) => {
   const { t } = useLocale();
+  const pathname = usePathname();
   const webappUrl = WEBAPP_URL.replace(/^https?:\/\//, "");
   const displayUrl = `${webappUrl}/${slug || ""}`;
+
+  // Animation variants for entry and exit
+  const containerVariants = {
+    initial: {
+      opacity: 0,
+      y: -20,
+    },
+    animate: {
+      opacity: 1,
+      y: 0,
+    },
+    exit: {
+      opacity: 0,
+      y: 20,
+    },
+  };
 
   const events: Array<{
     title: string;
@@ -60,7 +79,7 @@ export const OnboardingOrganizationBrowserView = ({
   ];
 
   return (
-    <div className="bg-default border-subtle hidden h-full w-full flex-col rounded-l-2xl border xl:flex">
+    <div className="bg-default border-subtle hidden h-full w-full flex-col rounded-l-2xl border-y border-s xl:flex overflow-hidden">
       {/* Browser header */}
       <div className="border-subtle flex min-w-0 shrink-0 items-center gap-3 rounded-t-2xl border-b bg-white p-3">
         {/* Navigation buttons */}
@@ -77,7 +96,18 @@ export const OnboardingOrganizationBrowserView = ({
       </div>
       {/* Content */}
       <div className="bg-muted h-full pl-11 pt-11">
-        <div className="bg-default border-muted flex h-full w-full flex-col overflow-hidden rounded-xl border">
+        <AnimatePresence mode="wait">
+          <motion.div
+            key={pathname}
+            className="bg-default border-muted flex h-full w-full flex-col overflow-hidden rounded-l-xl border-y border-s"
+            variants={containerVariants}
+            initial="initial"
+            animate="animate"
+            exit="exit"
+            transition={{
+              duration: 0.5,
+              ease: "backOut",
+            }}>
           {/* Organization Profile Header with Banner */}
           <div className="border-subtle flex flex-col border-b">
             <div className="relative">
@@ -140,7 +170,8 @@ export const OnboardingOrganizationBrowserView = ({
               </div>
             ))}
           </div>
-        </div>
+          </motion.div>
+        </AnimatePresence>
       </div>
     </div>
   );
